Update EntityIdentifier.md
Stephen, 
I am hoping we can add the ASIC ID to the data standard. The ASIC ID is a unique identifier for a financial adviser. It is public, published by ASIC and remains with the adviser forever - regardless of change of organisation or licensee. This is a powerful way to track adviser activity and identify distribution by that adviser across platforms and products. I would love to have a further discussion with you or the appropriate person. Please contact me at Lazard Asset Management - I won't put my contact details here, but I shouldn't be that hard to find! 
Warm Regards, 
Andrew Kempe
Head of Business Management
Lazard Asset Management